+++ src/SuperMarket.java
@@ -16,7 +16,7 @@ public class SuperMarket {
     // Na wat zoeken kwam ik er achter dat je de placeholders zo kan definiëren dat dezelfde waarde op verschillende
     // placeholders neergezet kan worden me de juiste syntax. In mijn implementatie krijgen beide %2$s definities dezelfde waarde.
     // De "s" (string), "d" (decimal integer) of "f" (float) duiden het type van de waarde in de placeholder aan.
-    private void buyItem(Product product, short amount) {
+    private void buyItem(Product product, int amount) {
         if (amount <= product.amount) {
             float totalPrice = amount * product.price;
             System.out.printf("You bought %d %s(s) for %f euro%n", amount, product.name, totalPrice);
@@ -26,19 +26,19 @@ public class SuperMarket {
         }
     }
 
-    public void buyBread(short amount) {
+    public void buyBread(int amount) {
         buyItem(bread, amount);
     }
 
-    public void buyFruit(short amount) {
+    public void buyFruit(int amount) {
         buyItem(fruit, amount);
     }
 
-    public void buyToiletPaper(short amount) {
+    public void buyToiletPaper(int amount) {
         buyItem(toiletPaper, amount);
     }
 
-    public void buyCheese(short amount) {
+    public void buyCheese(int amount) {
         buyItem(cheese, amount);
     }
 }
